Move is_order_preserving_cast and is_numeric_widening to arrow_tools::schema
