+++ crates/arrow_tools/src/schema.rs
@@ -315,6 +315,62 @@ fn normalize_dictionary_data_type(data_type: &DataType) -> DataType {
     }
 }
 
+/// Returns `true` if casting from `from` to `to` is guaranteed to preserve
+/// sort order (i.e. the cast is a monotonically non-decreasing function).
+///
+/// Covers:
+/// - Identity (same type)
+/// - Any temporal→temporal (Arrow stores UTC instants; timezone is metadata)
+/// - Safe numeric widening (same-sign widening, float widening, int→float
+///   where all source values are exactly representable)
+#[must_use]
+pub fn is_order_preserving_cast(from: &DataType, to: &DataType) -> bool {
+    if from == to {
+        return true;
+    }
+    if from.is_temporal() && to.is_temporal() {
+        return true;
+    }
+    if from.is_numeric() && to.is_numeric() {
+        return is_numeric_widening(from, to);
+    }
+    false
+}
+
+/// Same-signedness integer widening, float widening, and integer→float where
+/// all values are exactly representable.
+#[must_use]
+pub fn is_numeric_widening(from: &DataType, to: &DataType) -> bool {
+    matches!(
+        (from, to),
+        (
+            DataType::Int8,
+            DataType::Int16 | DataType::Int32 | DataType::Int64
+        ) | (DataType::Int16, DataType::Int32 | DataType::Int64)
+            | (DataType::Int32, DataType::Int64)
+            | (
+                DataType::UInt8,
+                DataType::UInt16 | DataType::UInt32 | DataType::UInt64
+            )
+            | (DataType::UInt16, DataType::UInt32 | DataType::UInt64)
+            | (DataType::UInt32, DataType::UInt64)
+            | (DataType::Float32, DataType::Float64)
+            | (
+                DataType::Int8
+                    | DataType::Int16
+                    | DataType::Int32
+                    | DataType::UInt8
+                    | DataType::UInt16
+                    | DataType::UInt32,
+                DataType::Float64
+            )
+            | (
+                DataType::Int8 | DataType::Int16 | DataType::UInt8 | DataType::UInt16,
+                DataType::Float32
+            )
+    )
+}
+
 pub fn set_computed_columns_meta<S: ::std::hash::BuildHasher>(
     schema: &mut Schema,
     computed_columns_meta: &HashMap<String, Vec<String>, S>,
@@ -1125,6 +1181,230 @@ mod tests {
             assert_eq!(col.value(0), "hello");
             assert!(col.is_null(1));
             assert_eq!(col.value(2), "world");
+        }
+    }
+
+    mod monotonic_cast_tests {
+        use super::super::{is_numeric_widening, is_order_preserving_cast};
+        use arrow::datatypes::{DataType, TimeUnit};
+
+        #[test]
+        fn timestamp_ns_to_us_same_tz() {
+            let from = DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()));
+            let to = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
+            assert!(is_order_preserving_cast(&from, &to));
+        }
+
+        #[test]
+        fn timestamp_s_to_ns() {
+            let from = DataType::Timestamp(TimeUnit::Second, Some("UTC".into()));
+            let to = DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()));
+            assert!(is_order_preserving_cast(&from, &to));
+        }
+
+        #[test]
+        fn timestamp_ms_to_us() {
+            let from = DataType::Timestamp(TimeUnit::Millisecond, Some("UTC".into()));
+            let to = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
+            assert!(is_order_preserving_cast(&from, &to));
+        }
+
+        #[test]
+        fn timestamp_different_tz_annotations() {
+            let from = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
+            let to = DataType::Timestamp(TimeUnit::Nanosecond, Some("America/New_York".into()));
+            assert!(is_order_preserving_cast(&from, &to));
+        }
+
+        #[test]
+        fn timestamp_tz_to_no_tz() {
+            let from = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
+            let to = DataType::Timestamp(TimeUnit::Nanosecond, None);
+            assert!(is_order_preserving_cast(&from, &to));
+        }
+
+        #[test]
+        fn date32_to_date64() {
+            assert!(is_order_preserving_cast(
+                &DataType::Date32,
+                &DataType::Date64
+            ));
+        }
+
+        #[test]
+        fn date64_to_date32() {
+            assert!(is_order_preserving_cast(
+                &DataType::Date64,
+                &DataType::Date32
+            ));
+        }
+
+        #[test]
+        fn int8_to_int16() {
+            assert!(is_numeric_widening(&DataType::Int8, &DataType::Int16));
+            assert!(is_order_preserving_cast(&DataType::Int8, &DataType::Int16));
+        }
+
+        #[test]
+        fn int8_to_int32() {
+            assert!(is_numeric_widening(&DataType::Int8, &DataType::Int32));
+        }
+
+        #[test]
+        fn int8_to_int64() {
+            assert!(is_numeric_widening(&DataType::Int8, &DataType::Int64));
+        }
+
+        #[test]
+        fn int16_to_int32() {
+            assert!(is_numeric_widening(&DataType::Int16, &DataType::Int32));
+        }
+
+        #[test]
+        fn int16_to_int64() {
+            assert!(is_numeric_widening(&DataType::Int16, &DataType::Int64));
+        }
+
+        #[test]
+        fn int32_to_int64() {
+            assert!(is_numeric_widening(&DataType::Int32, &DataType::Int64));
+        }
+
+        #[test]
+        fn uint8_to_uint16() {
+            assert!(is_numeric_widening(&DataType::UInt8, &DataType::UInt16));
+        }
+
+        #[test]
+        fn uint8_to_uint32() {
+            assert!(is_numeric_widening(&DataType::UInt8, &DataType::UInt32));
+        }
+
+        #[test]
+        fn uint8_to_uint64() {
+            assert!(is_numeric_widening(&DataType::UInt8, &DataType::UInt64));
+        }
+
+        #[test]
+        fn uint16_to_uint32() {
+            assert!(is_numeric_widening(&DataType::UInt16, &DataType::UInt32));
+        }
+
+        #[test]
+        fn uint16_to_uint64() {
+            assert!(is_numeric_widening(&DataType::UInt16, &DataType::UInt64));
+        }
+
+        #[test]
+        fn uint32_to_uint64() {
+            assert!(is_numeric_widening(&DataType::UInt32, &DataType::UInt64));
+        }
+
+        #[test]
+        fn float32_to_float64() {
+            assert!(is_numeric_widening(&DataType::Float32, &DataType::Float64));
+        }
+
+        #[test]
+        fn int32_to_float64() {
+            assert!(is_numeric_widening(&DataType::Int32, &DataType::Float64));
+        }
+
+        #[test]
+        fn uint32_to_float64() {
+            assert!(is_numeric_widening(&DataType::UInt32, &DataType::Float64));
+        }
+
+        #[test]
+        fn int16_to_float32() {
+            assert!(is_numeric_widening(&DataType::Int16, &DataType::Float32));
+        }
+
+        #[test]
+        fn uint16_to_float32() {
+            assert!(is_numeric_widening(&DataType::UInt16, &DataType::Float32));
+        }
+
+        #[test]
+        fn int8_to_float32() {
+            assert!(is_numeric_widening(&DataType::Int8, &DataType::Float32));
+        }
+
+        #[test]
+        fn int8_to_float64() {
+            assert!(is_numeric_widening(&DataType::Int8, &DataType::Float64));
+        }
+
+        #[test]
+        fn int32_to_uint32_not_monotonic() {
+            assert!(!is_numeric_widening(&DataType::Int32, &DataType::UInt32));
+            assert!(!is_order_preserving_cast(
+                &DataType::Int32,
+                &DataType::UInt32
+            ));
+        }
+
+        #[test]
+        fn uint32_to_int32_not_monotonic() {
+            assert!(!is_numeric_widening(&DataType::UInt32, &DataType::Int32));
+        }
+
+        #[test]
+        fn int64_to_int32_not_widening() {
+            assert!(!is_numeric_widening(&DataType::Int64, &DataType::Int32));
+        }
+
+        #[test]
+        fn float64_to_float32_not_widening() {
+            assert!(!is_numeric_widening(&DataType::Float64, &DataType::Float32));
+        }
+
+        #[test]
+        fn int64_to_float64_not_widening() {
+            assert!(!is_numeric_widening(&DataType::Int64, &DataType::Float64));
+        }
+
+        #[test]
+        fn uint64_to_float64_not_widening() {
+            assert!(!is_numeric_widening(&DataType::UInt64, &DataType::Float64));
+        }
+
+        #[test]
+        fn int32_to_float32_not_widening() {
+            assert!(!is_numeric_widening(&DataType::Int32, &DataType::Float32));
+        }
+
+        #[test]
+        fn uint32_to_float32_not_widening() {
+            assert!(!is_numeric_widening(&DataType::UInt32, &DataType::Float32));
+        }
+
+        #[test]
+        fn utf8_to_int64_not_order_preserving() {
+            assert!(!is_order_preserving_cast(&DataType::Utf8, &DataType::Int64));
+        }
+
+        #[test]
+        fn int64_to_utf8_not_order_preserving() {
+            assert!(!is_order_preserving_cast(&DataType::Int64, &DataType::Utf8));
+        }
+
+        #[test]
+        fn utf8_to_utf8_same_type() {
+            assert!(is_order_preserving_cast(&DataType::Utf8, &DataType::Utf8));
+        }
+
+        #[test]
+        fn same_type_always_preserving() {
+            assert!(is_order_preserving_cast(&DataType::Int64, &DataType::Int64));
+            assert!(is_order_preserving_cast(
+                &DataType::Float64,
+                &DataType::Float64
+            ));
+            assert!(is_order_preserving_cast(
+                &DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into())),
+                &DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()))
+            ));
         }
     }
 }
