+++ crates/runtime-datafusion/src/execution_plan/schema_cast.rs
@@ -14,8 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use arrow::datatypes::{DataType, Field, Schema, SchemaRef};
+use arrow::datatypes::{Field, Schema, SchemaRef};
 use arrow_tools::record_batch;
+use arrow_tools::schema::is_order_preserving_cast;
 use async_stream::stream;
 use async_trait::async_trait;
 use datafusion::catalog::Session;
@@ -45,42 +46,6 @@ use std::any::Any;
 use std::clone::Clone;
 use std::fmt;
 use std::sync::Arc;
-
-/// Returns `true` if casting from `from` to `to` is guaranteed to preserve ordering.
-///
-/// Mirrors DataFusion's `CastExpr::get_properties()`: temporal→temporal and
-/// numeric widening casts are monotonic. Arrow timestamps are UTC instants
-/// regardless of timezone annotation, so unit conversions are always monotonic.
-fn is_order_preserving_cast(from: &DataType, to: &DataType) -> bool {
-    if from == to {
-        return true;
-    }
-    if from.is_temporal() && to.is_temporal() {
-        return true;
-    }
-    if from.is_numeric() && to.is_numeric() {
-        return is_numeric_widening(from, to);
-    }
-    false
-}
-
-/// Same-signedness integer widening, float widening, and integer→float where
-/// all values are exactly representable.
-fn is_numeric_widening(from: &DataType, to: &DataType) -> bool {
-    use DataType::*;
-    matches!(
-        (from, to),
-        (Int8, Int16 | Int32 | Int64)
-            | (Int16, Int32 | Int64)
-            | (Int32, Int64)
-            | (UInt8, UInt16 | UInt32 | UInt64)
-            | (UInt16, UInt32 | UInt64)
-            | (UInt32, UInt64)
-            | (Float32, Float64)
-            | (Int8 | Int16 | Int32 | UInt8 | UInt16 | UInt32, Float64)
-            | (Int8 | Int16 | UInt8 | UInt16, Float32)
-    )
-}
 
 pub struct SchemaCastScanExec {
     input: Arc<dyn ExecutionPlan>,
@@ -428,7 +393,7 @@ impl TableProvider for EnsureSchema {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use arrow::datatypes::{DataType, Field, Schema};
+    use arrow::datatypes::{DataType, Field, Schema, TimeUnit};
     use datafusion::physical_expr::LexOrdering;
     use datafusion::physical_plan::empty::EmptyExec;
     use datafusion::physical_plan::expressions::col as physical_col;
@@ -843,250 +808,5 @@ mod tests {
             "Column 'a' should be remapped to index 1 in the output schema"
         );
         assert_eq!(col.name(), "a");
-    }
-
-    // ─── is_order_preserving_cast / is_numeric_widening unit tests ───
-
-    mod monotonic_cast_tests {
-        use super::super::{is_numeric_widening, is_order_preserving_cast};
-        use arrow::datatypes::{DataType, TimeUnit};
-
-        // --- Temporal casts (all should be order-preserving) ---
-
-        #[test]
-        fn timestamp_us_to_ns_same_tz() {
-            let from = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
-            let to = DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()));
-            assert!(is_order_preserving_cast(&from, &to));
-        }
-
-        #[test]
-        fn timestamp_ns_to_us_same_tz() {
-            let from = DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()));
-            let to = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
-            assert!(is_order_preserving_cast(&from, &to));
-        }
-
-        #[test]
-        fn timestamp_s_to_ns() {
-            let from = DataType::Timestamp(TimeUnit::Second, Some("UTC".into()));
-            let to = DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()));
-            assert!(is_order_preserving_cast(&from, &to));
-        }
-
-        #[test]
-        fn timestamp_ms_to_us() {
-            let from = DataType::Timestamp(TimeUnit::Millisecond, Some("UTC".into()));
-            let to = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
-            assert!(is_order_preserving_cast(&from, &to));
-        }
-
-        #[test]
-        fn timestamp_different_tz_annotations() {
-            // Arrow timestamps are always UTC instants; tz is display metadata.
-            let from = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
-            let to = DataType::Timestamp(TimeUnit::Nanosecond, Some("America/New_York".into()));
-            assert!(is_order_preserving_cast(&from, &to));
-        }
-
-        #[test]
-        fn timestamp_tz_to_no_tz() {
-            let from = DataType::Timestamp(TimeUnit::Microsecond, Some("UTC".into()));
-            let to = DataType::Timestamp(TimeUnit::Nanosecond, None);
-            assert!(is_order_preserving_cast(&from, &to));
-        }
-
-        #[test]
-        fn date32_to_date64() {
-            assert!(is_order_preserving_cast(
-                &DataType::Date32,
-                &DataType::Date64
-            ));
-        }
-
-        #[test]
-        fn date64_to_date32() {
-            assert!(is_order_preserving_cast(
-                &DataType::Date64,
-                &DataType::Date32
-            ));
-        }
-
-        // --- Numeric widening (should be order-preserving) ---
-
-        #[test]
-        fn int8_to_int16() {
-            assert!(is_numeric_widening(&DataType::Int8, &DataType::Int16));
-            assert!(is_order_preserving_cast(&DataType::Int8, &DataType::Int16));
-        }
-
-        #[test]
-        fn int8_to_int32() {
-            assert!(is_numeric_widening(&DataType::Int8, &DataType::Int32));
-        }
-
-        #[test]
-        fn int8_to_int64() {
-            assert!(is_numeric_widening(&DataType::Int8, &DataType::Int64));
-        }
-
-        #[test]
-        fn int16_to_int32() {
-            assert!(is_numeric_widening(&DataType::Int16, &DataType::Int32));
-        }
-
-        #[test]
-        fn int16_to_int64() {
-            assert!(is_numeric_widening(&DataType::Int16, &DataType::Int64));
-        }
-
-        #[test]
-        fn int32_to_int64() {
-            assert!(is_numeric_widening(&DataType::Int32, &DataType::Int64));
-        }
-
-        #[test]
-        fn uint8_to_uint16() {
-            assert!(is_numeric_widening(&DataType::UInt8, &DataType::UInt16));
-        }
-
-        #[test]
-        fn uint8_to_uint32() {
-            assert!(is_numeric_widening(&DataType::UInt8, &DataType::UInt32));
-        }
-
-        #[test]
-        fn uint8_to_uint64() {
-            assert!(is_numeric_widening(&DataType::UInt8, &DataType::UInt64));
-        }
-
-        #[test]
-        fn uint16_to_uint32() {
-            assert!(is_numeric_widening(&DataType::UInt16, &DataType::UInt32));
-        }
-
-        #[test]
-        fn uint16_to_uint64() {
-            assert!(is_numeric_widening(&DataType::UInt16, &DataType::UInt64));
-        }
-
-        #[test]
-        fn uint32_to_uint64() {
-            assert!(is_numeric_widening(&DataType::UInt32, &DataType::UInt64));
-        }
-
-        #[test]
-        fn float32_to_float64() {
-            assert!(is_numeric_widening(&DataType::Float32, &DataType::Float64));
-        }
-
-        #[test]
-        fn int32_to_float64() {
-            assert!(is_numeric_widening(&DataType::Int32, &DataType::Float64));
-        }
-
-        #[test]
-        fn uint32_to_float64() {
-            assert!(is_numeric_widening(&DataType::UInt32, &DataType::Float64));
-        }
-
-        #[test]
-        fn int16_to_float32() {
-            assert!(is_numeric_widening(&DataType::Int16, &DataType::Float32));
-        }
-
-        #[test]
-        fn uint16_to_float32() {
-            assert!(is_numeric_widening(&DataType::UInt16, &DataType::Float32));
-        }
-
-        #[test]
-        fn int8_to_float32() {
-            assert!(is_numeric_widening(&DataType::Int8, &DataType::Float32));
-        }
-
-        #[test]
-        fn int8_to_float64() {
-            assert!(is_numeric_widening(&DataType::Int8, &DataType::Float64));
-        }
-
-        // --- Non-monotonic / disallowed casts ---
-
-        #[test]
-        fn int32_to_uint32_not_monotonic() {
-            // Signed → unsigned reinterprets negatives as large positives.
-            assert!(!is_numeric_widening(&DataType::Int32, &DataType::UInt32));
-            assert!(!is_order_preserving_cast(
-                &DataType::Int32,
-                &DataType::UInt32
-            ));
-        }
-
-        #[test]
-        fn uint32_to_int32_not_monotonic() {
-            // Unsigned → signed of same width: values > i32::MAX wrap.
-            assert!(!is_numeric_widening(&DataType::UInt32, &DataType::Int32));
-        }
-
-        #[test]
-        fn int64_to_int32_not_widening() {
-            // Narrowing: potential overflow changes ordering.
-            assert!(!is_numeric_widening(&DataType::Int64, &DataType::Int32));
-        }
-
-        #[test]
-        fn float64_to_float32_not_widening() {
-            assert!(!is_numeric_widening(&DataType::Float64, &DataType::Float32));
-        }
-
-        #[test]
-        fn int64_to_float64_not_widening() {
-            // Int64 values > 2^53 lose precision in Float64, which can reorder.
-            assert!(!is_numeric_widening(&DataType::Int64, &DataType::Float64));
-        }
-
-        #[test]
-        fn uint64_to_float64_not_widening() {
-            assert!(!is_numeric_widening(&DataType::UInt64, &DataType::Float64));
-        }
-
-        #[test]
-        fn int32_to_float32_not_widening() {
-            // Int32 values > 2^24 lose precision in Float32.
-            assert!(!is_numeric_widening(&DataType::Int32, &DataType::Float32));
-        }
-
-        #[test]
-        fn uint32_to_float32_not_widening() {
-            assert!(!is_numeric_widening(&DataType::UInt32, &DataType::Float32));
-        }
-
-        #[test]
-        fn utf8_to_int64_not_order_preserving() {
-            assert!(!is_order_preserving_cast(&DataType::Utf8, &DataType::Int64));
-        }
-
-        #[test]
-        fn int64_to_utf8_not_order_preserving() {
-            assert!(!is_order_preserving_cast(&DataType::Int64, &DataType::Utf8));
-        }
-
-        #[test]
-        fn utf8_to_utf8_same_type() {
-            assert!(is_order_preserving_cast(&DataType::Utf8, &DataType::Utf8));
-        }
-
-        #[test]
-        fn same_type_always_preserving() {
-            assert!(is_order_preserving_cast(&DataType::Int64, &DataType::Int64));
-            assert!(is_order_preserving_cast(
-                &DataType::Float64,
-                &DataType::Float64
-            ));
-            assert!(is_order_preserving_cast(
-                &DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into())),
-                &DataType::Timestamp(TimeUnit::Nanosecond, Some("UTC".into()))
-            ));
-        }
     }
 }
